refactor(chains): inline derived.rs and l1_registry.rs modules
Delete the two thin wrapper modules and move their logic into their natural homes: ChainConfig conversions into config.rs, Upgrades impl for RollupConfig into upgrades.rs, and the L1_CONFIGS static into ethereum/config.rs as a public export. All callers are updated to use L1_CONFIGS.get() directly, and tracing log calls are fixed to use structured key=value fields per codebase conventions.
